+++ contracts/zero-ex/test/wrappers.ts
@@ -82,7 +82,6 @@ export * from '../test/generated-wrappers/meta_transactions_feature';
 export * from '../test/generated-wrappers/mixin_balancer';
 export * from '../test/generated-wrappers/mixin_balancer_v2';
 export * from '../test/generated-wrappers/mixin_bancor';
-export * from '../test/generated-wrappers/mixin_clipper';
 export * from '../test/generated-wrappers/mixin_co_fi_x';
 export * from '../test/generated-wrappers/mixin_crypto_com';
 export * from '../test/generated-wrappers/mixin_curve';