+++ packages/asset-swapper/src/utils/market_operation_utils/sampler_operations.ts
@@ -20,7 +20,6 @@ import {
 import {
     BALANCER_V2_VAULT_ADDRESS_BY_CHAIN,
     BANCOR_REGISTRY_BY_CHAIN_ID,
-    CLIPPER_INFO_BY_CHAIN,
     DODOV1_CONFIG_BY_CHAIN_ID,
     DODOV2_FACTORIES_BY_CHAIN_ID,
     KYBER_CONFIG_BY_CHAIN_ID,
@@ -364,7 +363,7 @@ export class SamplerOperations {
         return new SamplerContractOperation({
             source,
             fillData: {
-                poolAddress: '0xe3a207e4225d459095491ea75d30b31968dff887',
+                poolAddress: providerAddress,
                 gasCost,
             },
             contract: this._samplerContract,
@@ -1258,29 +1257,18 @@ export class SamplerOperations {
                             this.getShellSellQuotes(pool, makerToken, takerToken, takerFillAmounts, source),
                         );
                     case ERC20BridgeSource.LiquidityProvider:
-                        const { poolAddress: clipperPoolAddress, tokens: clipperTokens } = CLIPPER_INFO_BY_CHAIN[
-                            this.chainId
-                        ];
-                        if (
-                            clipperPoolAddress === NULL_ADDRESS ||
-                            !clipperTokens.includes(makerToken) ||
-                            !clipperTokens.includes(takerToken)
-                        ) {
-                            return [];
-                        }
-                        // Clipper requires WETH to be represented as address(0)
-                        const adjustedMakerToken =
-                            makerToken === NATIVE_FEE_TOKEN_BY_CHAIN_ID[this.chainId] ? NULL_ADDRESS : makerToken;
-                        const adjustedTakerToken =
-                            takerToken === NATIVE_FEE_TOKEN_BY_CHAIN_ID[this.chainId] ? NULL_ADDRESS : takerToken;
-                        // Supports the PLP interface
-                        return this.getLiquidityProviderSellQuotes(
-                            clipperPoolAddress,
-                            adjustedMakerToken,
-                            adjustedTakerToken,
-                            takerFillAmounts,
-                            // tslint:disable-next-line: custom-no-magic-numbers
-                            0, // Not used for Clipper
+                        return getLiquidityProvidersForPair(	
+                            this.liquidityProviderRegistry,	
+                            takerToken,	
+                            makerToken,	
+                        ).map(({ providerAddress, gasCost }) =>	
+                            this.getLiquidityProviderSellQuotes(	
+                                providerAddress,	
+                                makerToken,	
+                                takerToken,	
+                                takerFillAmounts,	
+                                gasCost,	
+                            ),
                         );
                     case ERC20BridgeSource.MStable:
                         return getShellLikeInfosForPair(this.chainId, takerToken, makerToken, source).map(pool =>
@@ -1427,8 +1415,6 @@ export class SamplerOperations {
 
                         return this.getLidoSellQuotes(lidoInfo, makerToken, takerToken, takerFillAmounts);
                     }
-                    case ERC20BridgeSource.Clipper:
-                        return [];
                     default:
                         throw new Error(`Unsupported sell sample source: ${source}`);
                 }
@@ -1540,29 +1526,18 @@ export class SamplerOperations {
                             this.getShellBuyQuotes(pool, makerToken, takerToken, makerFillAmounts, source),
                         );
                     case ERC20BridgeSource.LiquidityProvider:
-                        const { poolAddress: clipperPoolAddress, tokens: clipperTokens } = CLIPPER_INFO_BY_CHAIN[
-                            this.chainId
-                        ];
-                        if (
-                            clipperPoolAddress === NULL_ADDRESS ||
-                            !clipperTokens.includes(makerToken) ||
-                            !clipperTokens.includes(takerToken)
-                        ) {
-                            return [];
-                        }
-                        // Clipper requires WETH to be represented as address(0)
-                        const adjustedMakerToken =
-                            makerToken === NATIVE_FEE_TOKEN_BY_CHAIN_ID[this.chainId] ? NULL_ADDRESS : makerToken;
-                        const adjustedTakerToken =
-                            takerToken === NATIVE_FEE_TOKEN_BY_CHAIN_ID[this.chainId] ? NULL_ADDRESS : takerToken;
-                        // Supports the PLP interface
-                        return this.getLiquidityProviderBuyQuotes(
-                            clipperPoolAddress,
-                            adjustedMakerToken,
-                            adjustedTakerToken,
-                            makerFillAmounts,
-                            // tslint:disable-next-line: custom-no-magic-numbers
-                            0, // Not used for Clipper
+                        return getLiquidityProvidersForPair(	
+                            this.liquidityProviderRegistry,	
+                            takerToken,	
+                            makerToken,	
+                        ).map(({ providerAddress, gasCost }) =>	
+                            this.getLiquidityProviderBuyQuotes(	
+                                providerAddress,	
+                                makerToken,	
+                                takerToken,	
+                                makerFillAmounts,	
+                                gasCost,	
+                            ),
                         );
                     case ERC20BridgeSource.MStable:
                         return getShellLikeInfosForPair(this.chainId, takerToken, makerToken, source).map(pool =>
@@ -1704,8 +1679,6 @@ export class SamplerOperations {
 
                         return this.getLidoBuyQuotes(lidoInfo, makerToken, takerToken, makerFillAmounts);
                     }
-                    case ERC20BridgeSource.Clipper:
-                        return [];
                     default:
                         throw new Error(`Unsupported buy sample source: ${source}`);
                 }