+++ packages/asset-swapper/CHANGELOG.md
@@ -5,6 +5,22 @@ Edit the package's CHANGELOG.json file only.
 
 CHANGELOG
 
+## v16.29.1 - _October 4, 2021_
+
+    * Remove `Clipper` as a custom liquidity source (#335)
+
+## v16.29.0 - _October 4, 2021_
+
+    * Initial integration of neon-router (behind feature flag) (#295)
+
+## v16.28.0 - _September 29, 2021_
+
+    * Update ExchangeProxySwapQuoteConsumer for Multiplex V2 and friends (#282)
+
+## v16.27.5 - _Invalid date_
+
+    * Remove protocol fees by setting `PROTOCOL_FEE_MULTIPLIER` to 0 (#333)
+
 ## v16.27.4 - _September 15, 2021_
 
     * Dependencies updated