+++ packages/protocol-utils/CHANGELOG.md
@@ -5,6 +5,14 @@ Edit the package's CHANGELOG.json file only.
 
 CHANGELOG
 
+## v1.9.2 - _October 4, 2021_
+
+    * Dependencies updated
+
+## v1.9.1 - _September 29, 2021_
+
+    * Dependencies updated
+
 ## v1.9.0 - _September 15, 2021_
 
     * Add 'TreasuryVote' class